// src/models/FieldType.ts
/**
 * Enum for supported field types in the form.
 */
export enum FieldType {
    TEXT = 'text',
    EMAIL = 'email',
    PASSWORD = 'password',
    NUMBER = 'number',
    TEXTAREA = 'textarea',
    SELECT = 'select',
    CHECKBOX = 'checkbox',
    RADIO = 'radio',
    DATE = 'date',
    TIME = 'time',
    FILE = 'file',
    PHONE = 'tel',
    URL = 'url',
    COLOR = 'color',
    RANGE = 'range',
    SEARCH = 'search',
    DATETIME = 'datetime-local',
    HIDDEN = 'hidden'
}