// src/models/FormLayoutTypes.ts
export enum FieldLayout {
  OUTLINED = 'outlined',
  FILLED = 'filled',
  UNDERLINED = 'underlined',
  FLOATING = 'floating',
  STANDARD = 'standard',
  MATERIAL = 'material'
}

export enum ButtonAlignment {
  LEFT = 'left',
  CENTER = 'center',
  RIGHT = 'right'
}

export enum FormVariant {
  LIGHT = 'light',
  DARK = 'dark'
}

export enum FormDensity {
  COMFORTABLE = 'comfortable',
  COMPACT = 'compact',
  STANDARD = 'standard'
}

export enum FormRoundness {
  NONE = 'none',
  SLIGHT = 'slight',
  MEDIUM = 'medium',
  FULL = 'full'
}